backend/requirements.txt
Django==4.2.22
django-cors-headers==4.3.1
django-filter==23.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
djoser==2.2.1
drf-spectacular==0.27.0
Pillow==10.4.0
psycopg2-binary==2.9.7
python-dotenv==0.19.0